v0.0.1

+++ delivery/sql/models/delivery.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Enum, Float, Integer, String
+from sqlalchemy import Boolean, Column, Enum, Numeric, String
 
 from delivery.models.common import AvailableCitiesEnum
 from delivery.sql.models.base import BaseModel
@@ -10,11 +10,11 @@ class Delivery(BaseModel):
     name = Column(String(255))
     in_ = Column(Enum(AvailableCitiesEnum))
     out = Column(Enum(AvailableCitiesEnum))
-    price = Column(Integer)
-    min_price = Column(Integer)
-    weight = Column(Float)
-    volume = Column(Float)
-    delivery_date = Column(Integer)
+    price = Column(Numeric)
+    min_price = Column(Numeric)
+    weight = Column(Numeric)
+    volume = Column(Numeric)
+    delivery_date = Column(Numeric)
     less = Column(Boolean)
     weight_multiplier = Column(Boolean)
     volume_multiplier = Column(Boolean)
